ビルドディレクトリがない場合にエラーになるのを修正

+++ scripts/build.sh
@@ -6,7 +6,9 @@ TRIM_SRC=./build/qujisho.trim.txt
 RELEASE_GENERAL=${BUILD_DIR}/qujisho-general.txt
 RELEASE_WIN=${BUILD_DIR}/qujisho-windows.txt
 
-rm -r $BUILD_DIR
+if [ -e $BUILD_DIR ]; then
+    rm -r $BUILD_DIR
+fi
 mkdir $BUILD_DIR
 
 # コメント行を取り除く。
